May 11 - convert $$ to integer and CSS cleanup (#11)
* dollars to integers

* cart CSS cleanup

+++ server/migrations/202304-005-create-order-items-table.js
@@ -30,11 +30,11 @@ module.exports = {
       },
       cost: {
         allowNull: false,
-        type: Sequelize.DECIMAL(10,2)
+        type: Sequelize.INTEGER
       },
       tax: {
         allowNull: false,
-        type: Sequelize.DECIMAL(10,2)
+        type: Sequelize.INTEGER
       },
       order_item_status: {
         allowNull: false,
